+++ src/com/billpiel/sayid/core.clj
@@ -13,7 +13,7 @@
 
 (def version
   "The current version of sayid as a string."
-  "0.2.0")
+  "0.3.0")
 
 (def workspace
   "The active workspace. Used by default in any function prefixed `ws-`